fix: stabilize ci and refresh release docs

+++ src/cli.py
@@ -69,7 +69,7 @@ def _pct(values: list[int], ratio: float) -> int:
     if not values:
         return 0
     ordered = sorted(int(v) for v in values)
-    idx = int(round((len(ordered) - 1) * max(0.0, min(1.0, float(ratio)))))
+    idx = round((len(ordered) - 1) * max(0.0, min(1.0, float(ratio))))
     return int(ordered[idx])
